+++ src/main/java/ai/philterd/philter/model/CreateApiKeyRequest.java
@@ -1,0 +1,43 @@
+/*******************************************************************************
+ * Copyright 2026 Philterd, LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License.  You may obtain a copy
+ * of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ ******************************************************************************/
+package ai.philterd.philter.model;
+
+import com.google.gson.annotations.Expose;
+import com.google.gson.annotations.SerializedName;
+
+import java.util.List;
+
+/**
+ * A request to create an API key for a user through Philter's provisioning API.
+ *
+ * <p>The scopes are required and are not defaulted. They must be a subset of the scopes held by the
+ * key making the request: a key cannot grant a scope it does not itself carry.</p>
+ */
+public class CreateApiKeyRequest {
+
+    @Expose
+    @SerializedName("scopes")
+    private List<String> scopes;
+
+    public List<String> getScopes() {
+        return scopes;
+    }
+
+    public void setScopes(List<String> scopes) {
+        this.scopes = scopes;
+    }
+
+}