+++ src/main/java/ai/philterd/philter/model/CreateUserRequest.java
@@ -1,0 +1,65 @@
+/*******************************************************************************
+ * Copyright 2026 Philterd, LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License.  You may obtain a copy
+ * of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ ******************************************************************************/
+package ai.philterd.philter.model;
+
+import com.google.gson.annotations.Expose;
+import com.google.gson.annotations.SerializedName;
+
+/**
+ * A request to create a non-administrator user through Philter's provisioning API.
+ *
+ * <p>The role is not a parameter. The endpoint only ever creates a non-administrator; an
+ * administrator account is made in the dashboard.</p>
+ */
+public class CreateUserRequest {
+
+    @Expose
+    @SerializedName("username")
+    private String username;
+
+    @Expose
+    @SerializedName("email")
+    private String email;
+
+    @Expose
+    @SerializedName("password")
+    private String password;
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+
+}